requirements.txt
SQLAlchemy>=1.4.49,<2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
faker==25.2.0
psycopg2-binary==2.9.9
pandas
numpy==1.26.4
kafka-python